// elvitrina-frontend/src/app/core/models/store/store-category-type.enum.ts
export enum StoreCategoryType {
  HANDMADE_JEWELRY = 'HANDMADE_JEWELRY',
  POTTERY_CERAMICS = 'POTTERY_CERAMICS',
  TEXTILES_FABRICS = 'TEXTILES_FABRICS',
  ART_PAINTINGS = 'ART_PAINTINGS',
  HOME_DECOR = 'HOME_DECOR',
  CLOTHING_ACCESSORIES = 'CLOTHING_ACCESSORIES',
  ECO_FRIENDLY = 'ECO_FRIENDLY',
  LOCAL_FOODS = 'LOCAL_FOODS',
  HEALTH_WELLNESS = 'HEALTH_WELLNESS',
  BOOKS_STATIONERY = 'BOOKS_STATIONERY',
  TOYS_GAMES = 'TOYS_GAMES',
  VINTAGE_ANTIQUES = 'VINTAGE_ANTIQUES',
  DIGITAL_PRODUCTS = 'DIGITAL_PRODUCTS',
  CRAFTS_DIY = 'CRAFTS_DIY',
  PET_SUPPLIES = 'PET_SUPPLIES'
} 